lib/pull: Move logic for skipping scratch deltas
The rest of the logic for what delta to choose is in
`get_best_static_delta_start_for`, so move the logic for skipping
scratch deltas there, too. This makes the handling in `initiate_request`
more straightforward by returning no match when it should be skipped so
that it goes directly into the object pull path.