lib/pull: Only skip scratch deltas when normal commit is reachable
In e7305bbc8a262d595c7d1c507be7fba0e5a88d06 a heuristic was added that
if a scratch delta exists but you already have a commit on that ref that
an object pull should be preferred. The idea is that if you already have
something on the ref, then it's likely that an object pull require less
data to be fetched than pulling the entire new commit.

However, this doesn't take into account the commit state. If the
existing commit is partial, then it might be better to fetch the scratch
delta. On the other hand, if the existing commit is partial only because
a metadata only pull has been done, then the parent commit might be
normal and then you'd want to prefer an object pull again. Instead of
considering just the HEAD local commit, walk the parents and check if
any of them are normal commits. Only then use an object pull.