lib/pull: Skip scratch deltas if normal commit with same bindings found
If a normal commit cannot be found in the history of the current ref,
try to find one by looking at the collection and ref bindings in all the
commits in the repo. This will only work if the remote is using
collection IDs since otherwise just looking at the ref bindings in the
local commits would be ambiguous about what remote the commit came from.
If a normal commit with matching bindings is found, prefer an object
pull.

With this change it's possible to have a partial local HEAD commit (from
a previous metadata only pull, for example) and broken history to a
normal ancestor and correctly use the heuristic that an object pull
should be preferred. This is a common case since user repos are likely
to have missed intervening commits on the same remote ref.