tests: Remove (spurious) key generation to stdout
Nothing is using this right now, I don't quite understand
the intention of this part of the test.

And doing it to stdout will cause random test failures because
the key might happen to start with `ok` which will make TAP
think it's a passed test, and it's an error to have unplanned tests.