ot-fs-utils.c: try $TMPDIR if /var/tmp fails
In `ot_map_anonymous_tmpfile_from_content` we call `libglnx`'s
`glnx_open_anonymous_tmpfile` function, which attempts to create an fd
in `/var/tmp` regardless of the value of `$TMPDIR`.

This is _usually_ okay, but can fail in some contexts, such as in the
[NixOS][1] build environment, which doesn't have `/var` mapped at all.

To avoid failing in this case, if `glnx_open_anonymous_tmpfile` fails,
we retrieve the value of `$TMPDIR` manually and try calling
`glnx_open_anonymous_tmpfile_full` directly (which is equivalent save
for allowing us to choose the directory the fd is created in).

[1]: https://nixos.org/