commit: Optionally translate ima xattr
The security.ima extended attribute can only be written by root, but
there is the user.ima attribute that can be set by e.g. ima-evm-utils.
Allowing translation from user.ima to security.ima means that the tree
can be prepared with user.ima attributes, and after translation, stored
as security.ima in the tree.

This would result in a deployed security.ima attribute without needing
to run the tree compose as root.

Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>